Fixed integer conversion/format warning on x86_64 Linux
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@1731 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjnath/src/pjnath/ice_session.c
@@ -742,7 +742,7 @@ static const char *dump_check(char *buffer, unsigned bufsize,
     if (lcand->addr.addr.sa_family == pj_AF_INET()) {
 	len = pj_ansi_snprintf(buffer, bufsize,
 			       "%d: [%d] %s:%d-->%s:%d",
-			       GET_CHECK_ID(clist, check),
+			       (int)GET_CHECK_ID(clist, check),
 			       check->lcand->comp_id,
 			       laddr, (int)pj_ntohs(lcand->addr.ipv4.sin_port),
 			       pj_inet_ntoa(rcand->addr.ipv4.sin_addr),
